// libs/wizard/src/lib/wizard.enums.ts
export enum WizardOperator {
    AN = 'AN',
    OR = 'OR',
    GT = 'GT',
    GE = 'GE',
    EQ = 'EQ',
    NE = 'NE',
    LT = 'LT',
    LE = 'LE',
    COUNT = 'COUNT',
    IN = 'IN',
    NOT_IN = 'NOT_IN',
    INCLUDES = 'INCLUDES',
    NOT_INCLUDES = 'NOT_INCLUDES',
    INTERSECTS = 'INTERSECTS',
    NOT_INTERSECTS = 'NOT_INTERSECTS',
  }